// frontend/src/shared/api/generated/employers/models/industry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Employers
 * OpenAPI spec version: 0.1.0
 */

export type Industry = typeof Industry[keyof typeof Industry];


export const Industry = {
  fintech: 'fintech',
  ecommerce: 'ecommerce',
  govtech: 'govtech',
  telecom: 'telecom',
  healthtech: 'healthtech',
  edtech: 'edtech',
  gamedev: 'gamedev',
  media: 'media',
  logistics: 'logistics',
  industry: 'industry',
  energy: 'energy',
  cybersecurity: 'cybersecurity',
  ai: 'ai',
  travel: 'travel',
  real_estate: 'real_estate',
  outsource: 'outsource',
  other: 'other',
} as const;